Ignore untracked files for write-commit-push
If we add our default `.prettierrc.json` to the repository because the
original repository does not provide one, it will show up as untracked
file in `git status --porcelain`.

Since we don't want to trigger a commit when this untracked file is the
only reason for a dirty status, ignore untracked files when deciding
whether a commit is needed.